requirements.txt
openenv-core
fastapi
uvicorn[standard]
pydantic
numpy
openai
requests
Pillow